Refractor NEST Server with better error handling
